split registration task from cqrs integration tests
