+++ tests/BoltOn.Tests/Cqrs/TestCqrsRegistrationTask.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using BoltOn.Bootstrapping;
+using BoltOn.Data.EF;
+using Microsoft.EntityFrameworkCore;
+using BoltOn.Data;
+using Microsoft.EntityFrameworkCore.Diagnostics;
+
+namespace BoltOn.Tests.Cqrs
+{
+    public class TestCqrsRegistrationTask : IRegistrationTask
+    {
+        public void Run(RegistrationTaskContext context)
+        {
+            context.Container.AddDbContext<CqrsDbContext>(options =>
+            {
+                options.UseInMemoryDatabase("InMemoryDbCqrsDbContext");
+                options.ConfigureWarnings(x => x.Ignore(RelationalEventId.AmbientTransactionWarning));
+            });
+
+            context.Container.AddTransient<IRepository<TestCqrsWriteEntity>, CqrsRepository<TestCqrsWriteEntity, CqrsDbContext>>();
+            context.Container.AddTransient<IRepository<TestCqrsReadEntity>, CqrsRepository<TestCqrsReadEntity, CqrsDbContext>>();
+        }
+    }
+}
